+++ db_key_lab/android/app/src/main/kotlin/com/publicos/db_key_lab/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.publicos.db_key_lab
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()